% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geozl-package.R
\docType{package}
\name{geozl-package}
\alias{geozl}
\alias{geozl-package}
\title{geozl: geospatial codecs for OpenZL}
\description{
Compress raster tiles with GeoZL's spatial predictors, NoData mask and
bounded-error quantizers, and decode the frames any GeoZL reader writes.
Build a graph once with \code{\link[=geozl_graph]{geozl_graph()}}, compress tiles with
\code{\link[=geozl_compress]{geozl_compress()}}, and read them back with \code{\link[=geozl_decompress]{geozl_decompress()}}.
\code{\link[=geozl_profile]{geozl_profile()}} ranks recipes on a sample.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asterisk-labs/geozl}
  \item Report bugs at \url{https://github.com/asterisk-labs/geozl/issues}
}

}
\author{
\strong{Maintainer}: Cesar Aybar \email{cesar@asterisk.coop}

Other contributors:
\itemize{
  \item Asterisk Labs [copyright holder]
  \item Meta Platforms, Inc. and affiliates (OpenZL and Zstandard) [copyright holder]
  \item Yann Collet (LZ4) [copyright holder]
}

}
\keyword{internal}
