% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geozl-package.R
\docType{package}
\name{geozl-package}
\alias{geozl}
\alias{geozl-package}
\title{geozl: Geospatial Codec Extensions for 'OpenZL'}
\description{
R bindings to libgeozl, the geospatial codec extensions for OpenZL.
Build a graph once with \code{\link[=geozl_graph]{geozl_graph()}}, compress tiles with
\code{\link[=geozl_compress]{geozl_compress()}}, and read them back with \code{\link[=geozl_decompress]{geozl_decompress()}}.
\code{\link[=geozl_profile]{geozl_profile()}} ranks recipes on a sample.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asterisk-labs/geozl}
  \item Report bugs at \url{https://github.com/asterisk-labs/geozl/issues}
}

}
\author{
Cesar Aybar \email{cesar@asterisk.coop}
}
\keyword{internal}
